Extract Kore.Unification.SubstitutionSimplifier
